Update: move compareAddress
